scripts/pyproject.toml
[project]
name = "scripts"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "azure-identity>=1.25.1",
    "azure-storage-blob>=12.28.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
